fix: 4 Bugs behoben – TRV-Interval, Mold-ID, Add-Room-Typen, Coordinator-Int
- frontend: trv_min_send_interval im Edit-Room Save-Handler ergänzt (Wert
  wurde zwar im HTML angezeigt, aber beim Speichern nicht übergeben)
- frontend: HTML-ID m-mold-enabled → m-mold-protection im Add-Room Modal
  vereinheitlicht (ID-Mismatch zwischen Add- und Edit-Room Modal)
- frontend: Add-Room Save-Handler – parseFloat → parseInt für Sekundenfelder
  (window_reaction_time, window_close_delay, room_preheat_minutes) + alle
  direkten .value-Zugriffe auf ?.value mit Fallback-Werten abgesichert
- coordinator: float() → int() für CONF_WINDOW_REACTION_TIME und
  CONF_WINDOW_CLOSE_DELAY (konsistent mit __init__.py und const.py)

+++ custom_components/intelligent_heating_control/coordinator.py
@@ -1759,8 +1759,8 @@ class IHCCoordinator(DataUpdateCoordinator):
         - window_close_delay:   seconds after sensor goes OFF before IHC resumes heating (default 0 s)
         """
         room_id = room.get(CONF_ROOM_ID, "")
-        reaction_time = float(room.get(CONF_WINDOW_REACTION_TIME, DEFAULT_WINDOW_REACTION_TIME))
-        close_delay   = float(room.get(CONF_WINDOW_CLOSE_DELAY, DEFAULT_WINDOW_CLOSE_DELAY))
+        reaction_time = int(room.get(CONF_WINDOW_REACTION_TIME, DEFAULT_WINDOW_REACTION_TIME))
+        close_delay   = int(room.get(CONF_WINDOW_CLOSE_DELAY, DEFAULT_WINDOW_CLOSE_DELAY))
         now           = time.monotonic()
 
         # Check raw sensor state
